feat: new editor.enable-diagnostics option (#12203)
Sometimes diagnostics can get in the way, and we'd like to temporarily
hide them this PR adds a new editor.enable-diagnostics option. It
will hide obtrusive diagnostics, such as inline diagnostics or the
diagnostics which are overlaid on top of text. In the editor, it can be
toggled with :toggle enable-diagnostics. In the config file, it can be
set as follows:

```
[editor]
enable-diagnostics = false
```

It won't hide gutter symbols indicating that there are diagnostics
on lines, as those aren't obtrusive and it is useful to be aware that
there are diagnostics on certain lines. If we want to inspect what those
diagnostics are, we can just run :toggle enable-diagnostics again to
see them.

Closes #12154

Signed-off-by: Awiteb <a@4rs.nl>